Plotter.py is a watered down clone of Phobos
The tree function in question is setupTreeWidget().  You pass it a treewidget that the data will be tree'd out in, the dictionaries that contain the information you want to display (These should be the sub and fig dictionaries of the plotterated object), and the include list to show only the attributes that you want (for instance, Pie charts can't have a colorbar, so 'colorbar' is not included in the include list).