Updated Plotter
This is the new Phobos type thingy..... 
So, some caveats - 
the filters don't work as intended (Don't care though because they work at work).  
Still need to do something about saving figures (will work on this tonight)
Something is wrong while trying to import the Importerator version of Plotterator (plt is not defined and it also doesn't print the lines that    it is suppose to when it is importing)
Should Work other than that. Took out all dependencies on where I was keeping the file on my machine.  Will only work in the position that it is currently at in the REPO though.