Added BulkMinDistancePolygon O(n^2)
This function takes in either a list of polygons or a dictionary of labeled polygons.  Returns a dataframe with all of the relevant data needed to determine which polygons intersected and if they meet a specific threshold. This takes all unique combinations of polygons passed. Idk how Hamilton wants to use this, so I just went ahead and made this.